fix(product): Resolve images not appearing in related products
The 'products/category' API endpoint used for the related products section was outdated and no longer supported by the latest version of fakestoreapi.com.

This resulted in invalid image URLs that ended with '.jpg'. The fix ensures that valid image URLs, which now use the 't.png' extension as per the current API, are correctly handled, restoring the images in the related products section.